[chore #163054764] refactored meetup controller
